refactoring organization show view
