+++ frontend/src/resources/Organization/OrganizationShow/OrganizationShowEquipmentLayout.js
@@ -1,0 +1,37 @@
+import React from 'react';
+
+import { 
+  NumberField,
+  ReferenceField,
+  SimpleShowLayout,
+  TextField
+ } from 'react-admin';
+
+import useStyles from './OrganizationShowUseStyles';
+
+const OrganizationShowEquipmentLayout = ({...props}) => {
+  const classes = useStyles();
+  return (
+    <SimpleShowLayout>
+      <TextField source="pair:label" />
+      <ReferenceField source="petr:hasEquipmentType" reference="EquipmentType" link={false}>
+        <TextField source="pair:label" />
+      </ReferenceField>
+      <TextField source="pair:description" />
+      <TextField source="petr:model" />
+      <NumberField source="petr:amount" />
+      <ReferenceField source="petr:hasRate" reference="Rate" link={false}>
+        <TextField source="pair:label" />
+      </ReferenceField>
+      <TextField source="petr:availablity" />
+      <ReferenceField source="petr:hasAccessModality" reference="AccessModality" link={false}>
+        <TextField source="pair:label" />
+      </ReferenceField>
+      <ReferenceField source="pair:hasLocation" reference="Space" link={false}>
+        <TextField source="pair:label" />
+      </ReferenceField>
+    </SimpleShowLayout>
+  )
+}
+
+export default OrganizationShowEquipmentLayout;
